Henter landekoder fra backend
